#312 Add typed json response
