+++ source/index.ts
@@ -43,6 +43,7 @@ export {
 	AfterResponseHook,
 } from './types/hooks.js';
 
-export {ResponsePromise} from './types/response.js';
+export {ResponsePromise} from './types/responsePromise.js';
+export {KyResponse} from './types/response.js';
 export {HTTPError} from './errors/HTTPError.js';
 export {TimeoutError} from './errors/TimeoutError.js';
